Actualizado paso de datos a WebController

+++ VotingSystem_4b/src/main/java/es/uniovi/asw/Calculate/voters/VotersType.java
@@ -4,10 +4,10 @@ import java.util.Map;
 
 import es.uniovi.asw.service.GetVotes;
 
-
 public interface VotersType {
 
-	public Map<String,Double> getResult();
+	public Map<String, Integer> getResult();
+
 	public void actualize(GetVotes votes);
-	
+
 }
